Cleaned some code up
Forms still not showing on sidebar, and a new bug has been discovered:
Now that forms are used in selections (not groups), moving from a set
with a form to any other set (without the form) while the form is
selected causes a segfault.
This is because the form is in the selection, but you are moving to a
set where the form is out of scope.