Added form container struct
This will allow tracking of a parent form across the entire show. This
enables the drawing widget to draw forms on midsets (that's not done
yet).
This also uses the bsd queue library to implement a linear linked list
for the form container. I may standardize many of the current lists to
use this library.